requirements.txt
rumps
requests
python-dotenv 